+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ## [Unreleased]
 
+### Added
+- Added support for rumble in `XboxController` ([support#1024]).
+
 ### Changed
 - Allow single floating point value for brightness array ([support#1547]).
 
+[support#1024]: https://github.com/pybricks/support/issues/1024
 [support#1547]: https://github.com/pybricks/support/issues/1547
 
 ## [3.4.0] - 2024-03-11